Delay numpy import in setup.py
distutils does not offer good support for build-time dependencies,
despite the option of "setup_requires". This change proposes one
(admittedly hacky) method of delaying the import of numpy until after
the setup requirements have been processed. pycocotools having a
build-time dependency on numpy that is not automatically handled
by a python package manager makes the PythonAPI difficult to
properly include as a dependency in other projects.

Alternative solutions might include:
1. use of a pyproject.toml [PEP-518](https://www.python.org/dev/peps/pep-0518/)
   to specify the build system, but this likely poses challenges for Python 2.7
   support.
2. build_ext overloading as done in SciPy: https://github.com/scipy/scipy/blob/master/setup.py#L205
   however, this is likely to run into issues since Cython also
   customizes build_ext.

If this solution is deemed too hacky, it would be helpful to
add a clear README that specifies that numpy is a build-time
dependency.